+++ src/ArcadePointsBot/Errors/TwitchAPI.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AvaloniaApplication1.Errors
+namespace ArcadePointsBot.Errors
 {
     public static class TwitchAPI
     {
