+++ src/ArcadePointsBot/Interop/Windows/NativeMethods.cs
@@ -6,7 +6,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Runtime.InteropServices;
 
-namespace AvaloniaApplication1.Interop.Windows;
+namespace ArcadePointsBot.Interop.Windows;
 
 internal static class NativeMethods
 {
