+++ src/ArcadePointsBot/Views/MainWindow.axaml
@@ -1,19 +1,18 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:local="using:ArcadePointsBot"
-        xmlns:vm="using:AvaloniaApplication1.ViewModels"
-        xmlns:mm="using:AvaloniaApplication1.Models"
-        xmlns:app="using:AvaloniaApplication1"
+        xmlns:vm="using:ArcadePointsBot.ViewModels"
+        xmlns:mm="using:ArcadePointsBot.Models"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
         xmlns:ia="clr-namespace:Avalonia.Xaml.Interactions.Core;assembly=Avalonia.Xaml.Interactions"
         xmlns:system="using:System"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="AvaloniaApplication1.Views.MainWindow"
+        x:Class="ArcadePointsBot.Views.MainWindow"
         x:DataType="vm:MainWindowViewModel"
         Icon="/Assets/avalonia-logo.ico"
-        Title="AvaloniaApplication1"
+        Title="ArcadePointsBot"
         Design.DataContext="{x:Static local:DesignData.MainWindowViewModel}">
 
   <Grid RowDefinitions="*,*,*,45">
