+++ src/ArcadePointsBot/Views/RewardActionView.axaml.cs
@@ -2,7 +2,7 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace AvaloniaApplication1.Views;
+namespace ArcadePointsBot.Views;
 
 public partial class RewardActionView : UserControl
 {
